backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
requests==2.32.5
beautifulsoup4==4.14.2
lxml==6.0.2
python-dotenv==1.2.1
perplexityai==0.20.0